pyproject.toml
[project]
name = "simple-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "openai>=2.17.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.3",
]

[project.scripts]
simple-rag = "simple_rag:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
